user_service/requirements.txt
fastapi
uvicorn
aiohttp
prometheus_client
PyJWT